Organize ncurses code and header files
